// src/app/core/api/model/filter.ts
/**
 * vacancy ANCFCC
 * API documentation
 *
 * The version of the OpenAPI document: v1
 * Contact: abdellah.elbekkali@digitransform.co
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Filter { 
    id?: number;
    name?: string | null;
    label?: string | null;
    list?: Array<string> | null;
    count?: number;
}